Combine the CircusClient and DaemonClient in one
There was no real reason to keep these two classes separate so their
functionality has been merged into the DaemonClient class. This is
the gateway to any interaction with the daemon for a given profile
